+++ pkg/response/paginated_response.go
@@ -1,12 +1,12 @@
 package response
 
 type PaginationMeta struct {
-	Page          int  `json:"page"`
-	PageSize      int  `json:"page_size"`
-	TotalElements int  `json:"total_elements"`
-	TotalPages    int  `json:"total_pages"`
-	IsFirst       bool `json:"is_first"`
-	IsLast        bool `json:"is_last"`
+	Page          int   `json:"page"`
+	PageSize      int   `json:"page_size"`
+	TotalElements int64 `json:"total_elements"`
+	TotalPages    int   `json:"total_pages"`
+	IsFirst       bool  `json:"is_first"`
+	IsLast        bool  `json:"is_last"`
 }
 
 type PaginatedResponse struct {
